factored out Line and Point classes
